requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
cycler==0.10.0
decorator==4.4.2
future==0.18.2
google-api-core==1.23.0
google-auth==1.22.1
google-cloud-vision==2.0.0
googleapis-common-protos==1.52.0
grpcio==1.33.1
idna==2.10
imageio==2.9.0
kiwisolver==1.2.0
libcst==0.3.13
matplotlib==3.3.2
mypy-extensions==0.4.3
networkx==2.5
numpy==1.19.2
opencv-python==4.4.0.44
Pillow==8.0.1
proto-plus==1.11.0
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
pytesseract==0.3.6
python-dateutil==2.8.1
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
scikit-image==0.17.2
scipy==1.5.3
six==1.15.0
tifffile==2020.10.1
torch==1.6.0
torchvision==0.7.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.11
